pyproject.toml
[project]
name = "chatgpt-clone"
version = "0.1.0"
description = "Um clone simples do ChatGPT usando Django e OpenAI API."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.4",
    "markdown>=3.8.2",
    "openai>=1.95.0",
    "python-dotenv>=1.1.1",
]
